Created a separate launch script for trait mapping
